v1.0
The first version of RLStaller. It currently supports selecting 1 of 4 XBox controllers to read from, and allows the user to select any button on their XBox controller as the "stall" button. The user can also input their air steer left, air roll right, and jump keyboard bindings in Rocket League so RLStaller knows which sequence to perform when the stall button is pressed. The user's preferences are automatically saved so they don't need to be reentered each time RLStaller is launched.